requirements.txt
appdirs==1.4.3
asgiref==3.3.1
beautifulsoup4==4.9.3
bleach==3.2.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-bootstrap-form==3.4
django-crispy-forms==1.10.0
django-filter==2.4.0
django-markdown-deux==1.0.5
django-rest-swagger==2.2.0
django-rq==2.4.0
django-todo==2.4.9
djangorestframework==3.12.2
drf-yasg==1.20.0
email-reply-parser==0.5.12
extensions==0.4
future==0.18.2
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
lxml==4.6.2
Markdown==3.3.3
markdown2==2.3.10
MarkupSafe==1.1.1
msgpack==0.6.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
Pygments==2.7.3
pyparsing==2.4.6
PySocks==1.7.1
python-twitter==3.5
pytoml==0.1.21
pytz==2020.4
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rq==1.7.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.14.0
soupsieve==2.1
sqlparse==0.4.1
tweepy==3.9.0
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
